testFirebase/server/requirements.txt
Flask==1.1.2
Flask-Cors==3.0.9
Flask-JSON==0.3.4
Flask-RESTful==0.3.8
flask-restplus==0.13.0
Flask-SQLAlchemy==2.4.4
gunicorn==20.0.4
Jinja2==2.11.2
Pyrebase-t==1.0.28